moved suite logic into registry
